+++ src/com/github/rviehmann/aoc2019/Intcode.java
@@ -53,7 +53,7 @@ public class Intcode {
             switch (mode) {
                 case POSITION_MODE:
                     growIfNecessary(pc + numParam);
-                    address = (int) memory[pc + numParam];
+                    address = Math.toIntExact(memory[pc + numParam]);
                     growIfNecessary(address);
                     return memory[address];
                 case IMMEDIATE_MODE:
@@ -61,7 +61,7 @@ public class Intcode {
                     return memory[pc + numParam];
                 case RELATIVE_MODE:
                     growIfNecessary(pc + numParam);
-                    address = (int) (memory[pc + numParam] + relativeBase);
+                    address = Math.toIntExact((memory[pc + numParam] + relativeBase));
                     growIfNecessary(address);
                     return memory[address];
             }
@@ -73,7 +73,7 @@ public class Intcode {
             switch (mode) {
                 case POSITION_MODE:
                     growIfNecessary(pc + numParam);
-                    address = (int) memory[pc + numParam];
+                    address = Math.toIntExact(memory[pc + numParam]);
                     growIfNecessary(address);
                     memory[address] = value;
                     return;
@@ -81,7 +81,7 @@ public class Intcode {
                     throw new IllegalArgumentException("Immediate mode not allowed for output: position=" + pc + ", mode=" + mode);
                 case RELATIVE_MODE:
                     growIfNecessary(pc + numParam);
-                    address = (int) (memory[pc + numParam] + relativeBase);
+                    address = Math.toIntExact((memory[pc + numParam] + relativeBase));
                     growIfNecessary(address);
                     memory[address] = value;
                     return;
@@ -121,10 +121,10 @@ public class Intcode {
 
         while (true) {
             long memAtPc = memory.getRaw(pc);
-            int opcode = (int) (memAtPc % 100);
-            int param1Mode = (int) ((memAtPc / 100) % 10);
-            int param2Mode = (int) ((memAtPc / 1000) % 10);
-            int param3Mode = (int) ((memAtPc / 10000) % 10);
+            int opcode = Math.toIntExact(memAtPc % 100);
+            int param1Mode = Math.toIntExact((memAtPc / 100) % 10);
+            int param2Mode = Math.toIntExact((memAtPc / 1000) % 10);
+            int param3Mode = Math.toIntExact((memAtPc / 10000) % 10);
 
             if (!MODES.contains(param1Mode)) {
                 throw new IllegalArgumentException("Invalid param1Mode: param1Mode=" + param1Mode + ", position=" + pc);
@@ -173,7 +173,7 @@ public class Intcode {
                     input1 = memory.read(pc, 1, param1Mode, relativeBase);
                     input2 = memory.read(pc, 2, param2Mode, relativeBase);
                     if (input1 != 0) {
-                        pc = (int) input2;
+                        pc = Math.toIntExact(input2);
                     } else {
                         pc += 3;
                     }
@@ -183,7 +183,7 @@ public class Intcode {
                     input1 = memory.read(pc, 1, param1Mode, relativeBase);
                     input2 = memory.read(pc, 2, param2Mode, relativeBase);
                     if (input1 == 0) {
-                        pc = (int) input2;
+                        pc = Math.toIntExact(input2);
                     } else {
                         pc += 3;
                     }
@@ -207,7 +207,7 @@ public class Intcode {
 
                 case 9: // adjust relative base
                     input1 = memory.read(pc, 1, param1Mode, relativeBase);
-                    relativeBase += (int) input1;
+                    relativeBase += Math.toIntExact(input1);
                     pc += 2;
                     break;
